add support for animation json formatting outside gamelab
