+++ apps/Gruntfile.js
@@ -582,6 +582,8 @@ describe('entry tests', () => {
     'levels/editors/_craft':
       './src/sites/studio/pages/levels/editors/_craft.js',
     'levels/editors/_dsl': './src/sites/studio/pages/levels/editors/_dsl.js',
+    'levels/editors/fields/_animation':
+      './src/sites/studio/pages/levels/editors/fields/_animation.js',
     'levels/editors/fields/_blockly':
       './src/sites/studio/pages/levels/editors/fields/_blockly.js',
     'levels/editors/fields/_callouts':
