+++ dashboard/app/views/levels/editors/fields/_animation.html.haml
@@ -4,6 +4,7 @@
       padding: 0.2em;
       font-weight: bold;
     }
+  %script{src: webpack_asset_path('js/levels/editors/fields/_animation.js')}
 
 %h1.control-legend{data: {toggle: "collapse", target: "#animations"}}
   Animations
